fix: apply code review fixes, add system tests, and enforce clang-tidy in CI
Guard player death against re-processing after game over, clamp
speed/damage/angular_velocity in pattern JSON, and extract a shared
ability-hit helper used by ground slam, concussion shot, and melee.

Add Catch2 coverage for tile collision, input, movement, and pattern
loader error paths (60 -> 111 tests).

Add a CI lint job running clang-tidy with warnings-as-errors. Curate
.clang-tidy to match project conventions, bump spdlog to 1.15.3 for
clang 17+ consteval compatibility, and fix all remaining warnings
(uninitialized SDL out-params, widening index casts, silent catch,
sign comparisons, rule-of-five on Input, ranges algorithms).

+++ src/rendering/tilemap.cpp
@@ -72,7 +72,8 @@ bool Tilemap::is_solid(float x, float y, float w, float h) const {
 
     for (int gy = min_gy; gy <= max_gy; ++gy) {
         for (int gx = min_gx; gx <= max_gx; ++gx) {
-            if (collision_grid_[static_cast<size_t>(gy * grid_w_ + gx)]) {
+            if (collision_grid_[static_cast<size_t>(gy) * static_cast<size_t>(grid_w_) +
+                                static_cast<size_t>(gx)]) {
                 return true;
             }
         }
@@ -84,7 +85,8 @@ bool Tilemap::is_cell_solid(int grid_x, int grid_y) const {
     if (grid_x < 0 || grid_x >= grid_w_ || grid_y < 0 || grid_y >= grid_h_) {
         return false;
     }
-    return collision_grid_[static_cast<size_t>(grid_y * grid_w_ + grid_x)];
+    return collision_grid_[static_cast<size_t>(grid_y) * static_cast<size_t>(grid_w_) +
+                           static_cast<size_t>(grid_x)];
 }
 
 const SpawnPoint* Tilemap::find_spawn(const std::string& name) const {
